refactor: beans separation & clean TypeScript code
fix form's default values (they are null now)

+++ web/3/src/main/scala/com/itmo/mrdvd/dto/Dot.scala
@@ -1,3 +1,3 @@
 package com.itmo.mrdvd.dto
 
-case class Dot(X: Float, Y: Float, R: Float)
+case class Dot(X: Double, Y: Double, R: Double)
